카트 추가. 효석씨 코드 변경 은
productRepository 에 추가 한것 외에 , coffeebean_view.html 여기에 form 테그 와, quantity 증감값넣는 코드 main.js 에 그를 위한 코드 추가

+++ webproject/src/main/java/springboot/webproject/service/CartService.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface CartService {
 
     // 장바구니에 상품 추가
-    void addProductToCart(UsersDTO usersId, int prodNo, int quantity) throws Exception;
+    void addProductToCart(UsersDTO usersId, long prodNo, int quantity) throws Exception;
 
     // 장바구니에서 상품 주문 완료 처리
     void completeOrder(int cartNo) throws Exception;
